Best practice: Unit tests for Nummber Parser

+++ app/src/test/java/com/rtdti/calc16/NumberFormatTest.kt
@@ -6,7 +6,8 @@ import org.junit.Assert.*
 import org.junit.Test
 
 class NumberFormatTest {
-    fun twoto(e: Double) = Math.pow(2.0,e)
+    fun twoto(e: Int) = Math.pow(2.0, e.toDouble())
+
     @Test
     fun formatterFloat() {
         val format = NumberFormat.FLOAT
@@ -16,28 +17,29 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("1"), formatter.format(1.0, formatState))
         assertEquals(AnnotatedString("-1"), formatter.format(-1.0, formatState))
         assertEquals(AnnotatedString("1.5"), formatter.format(1.5, formatState))
-        assertEquals(AnnotatedString("0.0009765625"), formatter.format(twoto(-10.0), formatState))
-        assertEquals(AnnotatedString("0.9990234375"), formatter.format(1-twoto(-10.0), formatState))
+        assertEquals(AnnotatedString("0.0009765625"), formatter.format(twoto(-10), formatState))
+        assertEquals(AnnotatedString("0.9990234375"), formatter.format(1 - twoto(-10), formatState))
         assertEquals(AnnotatedString("3.141592653589793"), formatter.format(Math.PI, formatState))
-        assertEquals(AnnotatedString("1024"), formatter.format(twoto(10.0), formatState))
+        assertEquals(AnnotatedString("1024"), formatter.format(twoto(10), formatState))
         assertEquals(AnnotatedString("1000000"), formatter.format(1e+6, formatState))
-        assertEquals(AnnotatedString("1048576"), formatter.format(twoto(20.0), formatState))
-        assertEquals(AnnotatedString("1073741824"), formatter.format(twoto(30.0), formatState))
-        assertEquals(AnnotatedString("1099511627776"), formatter.format(twoto(40.0), formatState))
-        assertEquals(AnnotatedString("1125899906842624"), formatter.format(twoto(50.0), formatState))
-        // assertEquals(AnnotatedString("1152921504606846976"), formatter.format(twoto(60.0), formatState))
-        assertEquals(AnnotatedString("1.15292150460684698E18"), formatter.format(twoto(60.0), formatState))
-        assertEquals(AnnotatedString("1.1805916207174113E21"), formatter.format(twoto(70.0), formatState))
+        assertEquals(AnnotatedString("1048576"), formatter.format(twoto(20), formatState))
+        assertEquals(AnnotatedString("1073741824"), formatter.format(twoto(30), formatState))
+        assertEquals(AnnotatedString("1099511627776"), formatter.format(twoto(40), formatState))
+        assertEquals(AnnotatedString("1125899906842624"), formatter.format(twoto(50), formatState))
+        // assertEquals(AnnotatedString("1152921504606846976"), formatter.format(twoto(60), formatState))
+        assertEquals(AnnotatedString("1.15292150460684698E18"), formatter.format(twoto(60), formatState))
+        assertEquals(AnnotatedString("1.1805916207174113E21"), formatter.format(twoto(70), formatState))
         assertEquals(AnnotatedString("0.000001"), formatter.format(1e-6, formatState))
-        assertEquals(AnnotatedString("0.0009765625"), formatter.format(twoto(-10.0), formatState))
-        assertEquals(AnnotatedString("0.9990234375"), formatter.format(1-twoto(-10.0), formatState))
-        assertEquals(AnnotatedString("9.5367431640625E-7"), formatter.format(twoto(-20.0), formatState))
-        assertEquals(AnnotatedString("9.313225746154785E-10"), formatter.format(twoto(-30.0), formatState))
-        assertEquals(AnnotatedString("9.094947017729282E-13"), formatter.format(twoto(-40.0), formatState))
-        assertEquals(AnnotatedString("8.881784197001252E-16"), formatter.format(twoto(-50.0), formatState))
-        assertEquals(AnnotatedString("8.673617379884035E-19"), formatter.format(twoto(-60.0), formatState))
-        assertEquals(AnnotatedString("8.470329472543003E-22"), formatter.format(twoto(-70.0), formatState))
+        assertEquals(AnnotatedString("0.0009765625"), formatter.format(twoto(-10), formatState))
+        assertEquals(AnnotatedString("0.9990234375"), formatter.format(1 - twoto(-10), formatState))
+        assertEquals(AnnotatedString("9.5367431640625E-7"), formatter.format(twoto(-20), formatState))
+        assertEquals(AnnotatedString("9.313225746154785E-10"), formatter.format(twoto(-30), formatState))
+        assertEquals(AnnotatedString("9.094947017729282E-13"), formatter.format(twoto(-40), formatState))
+        assertEquals(AnnotatedString("8.881784197001252E-16"), formatter.format(twoto(-50), formatState))
+        assertEquals(AnnotatedString("8.673617379884035E-19"), formatter.format(twoto(-60), formatState))
+        assertEquals(AnnotatedString("8.470329472543003E-22"), formatter.format(twoto(-70), formatState))
     }
+
     @Test
     fun formatterHex() {
         val format = NumberFormat.HEX
@@ -50,6 +52,7 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("0x10"), formatter.format(16.0, formatState))
         assertEquals(AnnotatedString("0xfffffffffffffff0"), formatter.format(-16.0, formatState))
     }
+
     @Test
     fun formatterImproper() {
         val format = NumberFormat.IMPROPER
@@ -60,6 +63,7 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("-3 / 2"), formatter.format(-3/2.0, formatState))
         assertEquals(AnnotatedString("355 / 113 + ϵ"), formatter.format(Math.PI, formatState))
     }
+
     @Test
     fun formatterMixImperial() {
         val format = NumberFormat.MIXIMPERIAL
@@ -71,6 +75,7 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("3 - 145 / 1024 + ϵ"), formatter.format(Math.PI, formatState))
         assertEquals(AnnotatedString("5215 / 16384 + ϵ"), formatter.format(1/Math.PI, formatState))
     }
+
     @Test
     fun formatterFix() {
         val format = NumberFormat.FIX
@@ -82,6 +87,7 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("3.14"), formatter.format(Math.PI, formatState))
         assertEquals(AnnotatedString("0.32"), formatter.format(1/Math.PI, formatState))
     }
+
     @Test
     fun formatterSci() {
         val format = NumberFormat.SCI
@@ -93,6 +99,7 @@ class NumberFormatTest {
         assertEquals(AnnotatedString("3.14e+00"), formatter.format(Math.PI, formatState))
         assertEquals(AnnotatedString("3.18e-01"), formatter.format(1/Math.PI, formatState))
     }
+
     @Test
     fun formatterPrime() {
         val format = NumberFormat.PRIME
